+++ dotcom-rendering/src/components/SignInGate/gates/alternative-wording-guardian-live.tsx
@@ -1,7 +1,7 @@
 import { startPerformanceMeasure } from '@guardian/libs';
 import React, { Suspense } from 'react';
 import { Lazy } from '../../Lazy';
-import { canShowSignInGateWithOffers } from '../displayRule';
+import { canShowSignInGateWithOffers } from '../displayRules';
 import type { SignInGateComponent } from '../types';
 
 const SignInGateCustomizableText = React.lazy(() => {
